[RAIN-3803] Changed answers count type from Integer to Long

+++ core-services/egov-survey-services/src/main/java/org/egov/egovsurveyservices/web/models/SurveyEntity.java
@@ -59,7 +59,7 @@ public class SurveyEntity {
     private Boolean active;
 
     @JsonProperty("answersCount")
-    private Integer answersCount;
+    private Long answersCount;
 
     public SurveyEntity addQuestionsItem(Question questionItem) {
         if (this.questions == null) {
